feat(vr): changed render from browser to vr and have cube listen and change color based on words said
