--- browser/main.jsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import { render } from 'react-dom'
-import Root from './root'
-
-render(<div><h1>on Page!</h1><Root /></div>,
-  document.getElementById('main'))
