--- browser/style.css
@@ -1,54 +1,0 @@
-/*body, html {
-  margin: 0;
-}
-
-html {
-  height: 100%;
-  background-color: teal;
-}
-
-body {
-  height: inherit;
-  overflow: hidden;
-}
-
-h1, p {
-  font-family: sans-serif;
-  text-align: center;
-}
-
-div p {
-  padding: 20px;
-  background-color: rgba(0,0,0,0.2);
-}
-
-div {
-  overflow: auto;
-  position: absolute;
-  bottom: 0;
-  right: 0;
-  left: 0;
-}
-
-button {
-  margin: 0 auto;
-  display: block;
-  font-size: 1.1rem;
-  width: 170px;
-  line-height: 2;
-  margin-top: 30px;
-}
-
-@media all and (max-height: 410px) {
-  div {
-    position: static;
-  }
-}
-
-.phrase {
-  font-weight: bold;
-}
-
-.output {
-  font-style: italic;
-}*/
